chore: move spacing to another file
