+++ packages/styled/src/index.ts
@@ -17,12 +17,7 @@ import {
   noThemeProvider,
 } from './util'
 import { columnSystem } from '@charcoal-ui/foundation'
-import {
-  dur,
-  GradientDirection,
-  notDisabledSelector,
-  px,
-} from '@charcoal-ui/utils'
+import { dur, GradientDirection, px } from '@charcoal-ui/utils'
 import {
   Context,
   Internal,
@@ -32,6 +27,12 @@ import {
 } from './builders/internal'
 import { createColorCss, createGradientColorCss } from './builders/colors'
 import { createTypographyCss } from './builders/typography'
+import {
+  createSpacingCss,
+  spacingDirections,
+  spacingProperties,
+} from './builders/spacing'
+import { createOutlineColorCss, outlineType } from './builders/outline'
 export { type Modified, type ModifiedArgumented } from './builders/lib'
 export { default as TokenInjector } from './TokenInjector'
 export {
@@ -46,19 +47,6 @@ export {
 } from './helper'
 export { defineThemeVariables } from './util'
 export * from './SetThemeScript'
-
-const spacingProperties = ['margin', 'padding'] as const
-const spacingDirections = [
-  'top',
-  'right',
-  'bottom',
-  'left',
-  'vertical',
-  'horizontal',
-  'all',
-] as const
-type SpacingProperty = typeof spacingProperties[number]
-type SpacingDirection = typeof spacingDirections[number]
 
 const fixedProperties = ['width', 'height'] as const
 type FixedProperty = typeof fixedProperties[number]
@@ -220,82 +208,6 @@ function builder<T extends CharcoalAbstractTheme>(
     outlineObject
   )
 }
-
-function spacingProperty(
-  property: SpacingProperty,
-  direction: 'top' | 'right' | 'bottom' | 'left'
-) {
-  return `${property}-${direction}`
-}
-
-const createSpacingCss =
-  <T extends CharcoalAbstractTheme>(theme: { spacing: T['spacing'] }) =>
-  (
-    property: SpacingProperty,
-    modifiers: readonly [SpacingDirection, keyof T['spacing'] | 'auto'][]
-  ): Internal => {
-    const { top, right, bottom, left } = modifiers.reduce(
-      (acc, [direction, size]) => {
-        if (direction === 'all') {
-          acc.top = size
-          acc.right = size
-          acc.bottom = size
-          acc.left = size
-        } else if (direction === 'vertical') {
-          acc.top = size
-          acc.bottom = size
-        } else if (direction === 'horizontal') {
-          acc.right = size
-          acc.left = size
-        } else {
-          acc[direction] = size
-        }
-        return acc
-      },
-      {} as Partial<
-        Record<'top' | 'right' | 'bottom' | 'left', keyof T['spacing'] | 'auto'>
-      >
-    )
-
-    const hasVerticalPadding =
-      property === 'padding' &&
-      top !== undefined &&
-      bottom !== undefined &&
-      top !== 'auto' &&
-      bottom !== 'auto'
-
-    return internal(
-      ({ cancelHalfLeadingPx = 0 }) => ({
-        ...(top !== undefined && {
-          [spacingProperty(property, 'top')]:
-            top === 'auto'
-              ? 'auto'
-              : px(
-                  theme.spacing[top] +
-                    (hasVerticalPadding ? cancelHalfLeadingPx : 0)
-                ),
-        }),
-        ...(bottom !== undefined && {
-          [spacingProperty(property, 'bottom')]:
-            bottom === 'auto'
-              ? 'auto'
-              : px(
-                  theme.spacing[bottom] +
-                    (hasVerticalPadding ? cancelHalfLeadingPx : 0)
-                ),
-        }),
-        ...(right !== undefined && {
-          [spacingProperty(property, 'right')]:
-            right === 'auto' ? 'auto' : px(theme.spacing[right]),
-        }),
-        ...(left !== undefined && {
-          [spacingProperty(property, 'left')]:
-            left === 'auto' ? 'auto' : px(theme.spacing[left]),
-        }),
-      }),
-      hasVerticalPadding ? { hasVerticalPadding: true } : {}
-    )
-  }
 
 const createFixedPxCss =
   <T extends CharcoalAbstractTheme>(theme: T) =>
